Improves force torque sensor handling
Moves the `pose_in_base_from_name` helper function to the base `Controller` class to be accessible to all controllers.

Adds functionality to transform the force/torque sensor readings to the base frame for the `JointVelocityController`.
This allows for easier analysis and control of forces and torques in a consistent reference frame. Also compensates for gripper inertial properties when transforming force/torque readings.